refactor(code): refactor code with good library structure
