+++ tools/files.json
@@ -5,5 +5,6 @@
     { "name":"src" },
     { "name":"index.js", "remove":true },
     { "name":"index.js.map", "remove":true },
-    { "name":"index.d.ts", "remove":true }
+    { "name":"index.d.ts", "remove":true },
+    { "name":"lib", "remove":true }
 ]
